First working bruteforce positioner

+++ BattleshipGameboardGenerator/BoardCoordinate.cs
@@ -4,15 +4,15 @@ namespace BattleshipGameboardGenerator
 {
     public struct BoardCoordinate
     {
-        public BoardCoordinate(short row, short column)
+        public BoardCoordinate(int row, int column)
             : this()
         {
             Row = row;
             Column = column;
         }
 
-        public short Row { get; set; }
-        public short Column { get; set; }
+        public int Row { get; set; }
+        public int Column { get; set; }
         public override string ToString()
         {
             return ((char)((65) + Row)) + (Column + 1).ToString(CultureInfo.InvariantCulture);
